Core: Refactor `StringEscapeDecoder` into `coreutils_core::strings`
